Journey: Do not weaken hafas reference in stop items
The Stop constructor is called lazily and must have access to a hafas
reference for its add_message function and strptime_obj key. If the hafas
object goes out of scope before Journey->route is called, those may have
been garbage-collected already.

Apart from that, the weaken call was useless, since a Stop object does not
store the hafas reference provided in its constructor. So, as soon as
Journey->route is called, the hafas references in @{$journey->{route}} go out
of scope anyways.